Get rid of a Clist and move the functon to display floats with two decimals to gui_utils.
svn path=/trunk/; revision=27621

+++ gtk/gsm_map_stat.h
@@ -31,10 +31,10 @@
 /** Gsm map statistic data */
 typedef struct _gsm_map_stat_t {
     int			opr_code[GSM_MAP_MAX_NUM_OPR_CODES];
-    double		size[GSM_MAP_MAX_NUM_OPR_CODES];
+    int			size[GSM_MAP_MAX_NUM_OPR_CODES];
 
     int			opr_code_rr[GSM_MAP_MAX_NUM_OPR_CODES];
-    double		size_rr[GSM_MAP_MAX_NUM_OPR_CODES];
+    int			size_rr[GSM_MAP_MAX_NUM_OPR_CODES];
 } gsm_map_stat_t;
 
 /** Global gsm map statistic data */
